add user id on delete app

+++ internal/db/queries/queries.sql
@@ -8,7 +8,7 @@ SELECT * FROM applications WHERE status = $1 AND user_id = $2;
 SELECT * FROM applications WHERE id = $1;
 
 -- name: DeleteOneApplicationByID :one
-DELETE FROM applications WHERE id = $1 RETURNING *;
+DELETE FROM applications WHERE id = $1 AND user_id = $2 RETURNING *;
 
 -- name: CreateOneApplication :one
 INSERT INTO applications ( title_application, company, sent_date, status, notes, url_application, user_id ) VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *;
